requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9

# Job Search
google-search-results==2.4.2

# AI
google-generativeai==0.8.0

# Browser Automation
playwright==1.48.0

# Database
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
aiofiles==24.1.0
jinja2==3.1.4
